validaciones tarea

+++ app/Http/Controllers/TareaControlador.php
@@ -77,7 +77,7 @@ class TareaControlador extends Controller
             ]);
         }
         $proyectoGrupo = ProyectoGrupo::where('idProyecto', request('idProyecto'))->first();
-        
+
         if(UsuarioProyectoGrupo::where([['idProyectoGrupo', $proyectoGrupo->idProyectoGrupo],['idUsuario', Auth::id()]])->count() == 0){
             return response()->json([
                 'status' => 'ERROR',
@@ -118,7 +118,12 @@ class TareaControlador extends Controller
                 'result' => 'Inicia sesion para continuar'
             ]);
         }
-
+        if(AdministradorProyecto::where([['idProyecto', request('idProyecto')],['$idUsuario', Auth::id()]])->count() == 0){
+            return response()->json([
+                'status' => 'ERROR',
+                'result' => 'No tienes permiso para eliminar esta tarea'
+            ]);
+        }
         Tarea::where('idTarea', '=', request('idTarea'))->update(['estado' => 1]);
         $tarea = Tarea::where('idTarea', request('idTarea'))->first();
         $tareaProyectoGrupo = TareaProyectoGrupo::where('idTarea', $tarea->idTarea)->first();
@@ -139,7 +144,12 @@ class TareaControlador extends Controller
               'result' => 'Inicia sesion para continuar'
           ]);
         }
-
+        if(AdministradorProyecto::where([['idProyecto', request('idProyecto')],['$idUsuario', Auth::id()]])->count() == 0){
+            return response()->json([
+                'status' => 'ERROR',
+                'result' => 'No tienes permiso para asignar tareas en este proyecto'
+            ]);
+        }
         $tarea = Tarea::where('idTarea', '=', request('idTarea'))->first();
 
         if($tarea->estado != 2){
@@ -237,6 +247,17 @@ class TareaControlador extends Controller
     }
     public function subirEvidencia(){
       try{
+        $tareaProyectoGrupo = TareaProyectoGrupo::where('idTarea', request('idTarea'))->first();
+        $tareaUsuario = TareaUsuario::where('idTareaProyectoGrupo', $tareaProyectoGrupo->idTareaProyectoGrupo)->first();
+        $encargado = Usuario::where('idUsuario', $tareaUsuario->idUsuario)->first();
+
+        if($idUser != $encargado->idUsuario){
+            return response()->json([
+                'status' => 'ERROR',
+                'result' => 'No tienes permiso para subir evidencias en esta tarea'
+            ]);
+        }
+
         $tarea = Tarea::findOrFail(request('idTarea'));
 
         Storage::deleteDirectory('evidencias/'.request('idTarea'));
@@ -268,6 +289,20 @@ class TareaControlador extends Controller
     }
 
     public function descargarEvidencia($idTarea){
+      if(!Auth::check()){
+        return response()->json([
+            'status' => 'ERROR',
+            'result' => 'Inicia sesion para continuar',
+        ]);
+      }
+      $tareaProyectoGrupo::where('idTarea', $idTarea)->first();
+
+      if(UsuarioProyectoGrupo::where([['idProyectoGrupo', $tareaproyectoGrupo->idProyectoGrupo],['idUsuario', Auth::id()]])->count() == 0){
+          return response()->json([
+              'status' => 'ERROR',
+              'result' => 'No tienes permiso para descargar esta evidencia'
+          ]);
+      }
       try{
 
         $tarea = Tarea::findOrFail($idTarea);
@@ -291,8 +326,21 @@ class TareaControlador extends Controller
     }
 
     public function validarEvidencia(){
-        $id = 0;
-        if(Auth::check()) $id = Auth::id();
+        if(!Auth::check()){
+          return response()->json([
+              'status' => 'ERROR',
+              'result' => 'Inicia sesion para continuar',
+          ]);
+        }
+        $tareaProyectoGrupo::where('idTarea', request('idTarea'))->first();
+        $proyectoGrupo::where('idProyectoGrupo', $tareaProyectoGrupo->idProyectoGrupo)->first();
+        if(AdministradorProyecto::where([['idProyecto', $proyectoGrupo->idProyecto],['idUsuario', Auth::id()]])->count() == 0){
+            return response()->json([
+                'status' => 'ERROR',
+                'result' => 'No tienes permiso para validar esta evidencia',
+            ]);
+        }
+        $id = Auth::id();
         try{
             $tarea = Tarea::findOrFail(request('idTarea'));
             $tareaProyectoGrupo = TareaProyectoGrupo::where('idTarea', request('idTarea'))->first();
@@ -328,8 +376,21 @@ class TareaControlador extends Controller
 
     }
     public function rechazarEvidencia(){
-        $id = 0;
-        if(Auth::check()) $id = Auth::id();
+       if(!Auth::check()){
+          return response()->json([
+              'status' => 'ERROR',
+              'result' => 'Inicia sesion para continuar',
+          ]);
+       }
+       $tareaProyectoGrupo::where('idTarea', request('idTarea'))->first();
+       $proyectoGrupo::where('idProyectoGrupo', $tareaProyectoGrupo->idProyectoGrupo)->first();
+       if(AdministradorProyecto::where([['idProyecto', $proyectoGrupo->idProyecto],['idUsuario', Auth::id()]])->count() == 0){
+          return response()->json([
+              'status' => 'ERROR',
+              'result' => 'No tienes permiso para validar esta evidencia',
+          ]);
+       }
+       $id = Auth::id();
         try{
             $tarea = Tarea::findOrFail(request('idTarea'));
             $tareaProyectoGrupo = TareaProyectoGrupo::where('idTarea', request('idTarea'))->first();
